third_party/WebKit/LayoutTests/editing/shadow/shadow-selection-not-exported-expected.txt

CONSOLE WARNING: line 16: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS successfullyParsed is true

TEST COMPLETE


This test ensures shadow root won't expose to JS layer.

BEFORE
AFTER
